+++ src/main/java/am/nuaca/bim/repository/LanguageRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * @author Tigran Sargsyan on 23-May-20.
  */
-public interface LanguageRepository extends CrudRepository<Language, Long> {}
+public interface LanguageRepository extends CrudRepository<Language, Integer> {}
